requirements.txt
astroid==2.1.0
chardet==3.0.4
diff-match-patch==20121119
Django==2.0.2
django-csvimport==2.11
django-debug-toolbar==1.11
django-extensions==2.1.4
django-import-export==1.0.1
django-multiselectfield==0.0.3
django-pandas==0.5.1
et-xmlfile==1.0.1
isort==4.3.4
jdcal==1.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.3
odfpy==1.3.6
openpyxl==2.5.8
pandas==0.23.4
Pillow==5.2.0
psycopg2==2.7.5
pylint==2.2.2
python-dateutil==2.7.4
pytz==2018.5
PyYAML==3.13
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
typed-ast==1.1.0
unicodecsv==0.14.1
Werkzeug==0.14.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
